requirements.txt
mariadb==1.1.10
python-dotenv==1.0.1
requests==2.32.3
google-api-python-client==2.154.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
